YOLO_Streamlit/requirements.txt
streamlit
ultralytics
opencv-python-headless
Pillow
pandas
numpy
joblib
scikit-learn
matplotlib
seaborn